do not update locations for retroactive measures
Previously, new measurements predating existing measurements would override the
location of measurements with the same location.
This was a leftover from the previous implementation with location-references,
where this behaviour was desired. With the new implementation it cannot be
distinguished whether a measurement has an inferred or custom location.
For custom locations, the override would be an error!
For the sake of simplicity, only the simpler case is supported now
(-> no updating of inferred locations).